天気関連の処理をservicesに移動
