+++ backend/app/Http/Controllers/LocationController.php
@@ -5,16 +5,18 @@ namespace App\Http\Controllers;
 use App\Http\Requests\LocationRequest;
 use App\Location;
 use App\PostalCode;
+use App\Services\WeatherShowService;
 use App\User;
-use Datetime;
-use DateTimeZone;
 use Exception;
 use Illuminate\Http\Request;
 
 class LocationController extends Controller
 {
-    public function __construct()
+    protected $weatherShowService;
+
+    public function __construct(weatherShowService $weather_show_service)
     {
+        $this->weatherShowService = $weather_show_service;
         $this->authorizeResource(location::class, 'location');
     }
 
@@ -79,36 +81,10 @@ class LocationController extends Controller
     {
         // ユーザーの表示
         $user = User::where('id', $location->user_id)->first();
-        $weathers = [];
-        $temp_maxs = [];
-        $temp_mins = [];
-        $weather_icons = [];
-        $dates = [];
-        $times = [];
-        $tomorrow = new DateTime('+1 day');
-        $tomorrow = $tomorrow->format('Y年m月d日');
-        $zipcode = Location::zipcode($location->zipcode);
-
+        //天気の表示
         try {
-            // 例外が発生するおそれがあるコード
-            $response = Location::getWeather('forecast', $zipcode);
-
-            $weather_list = $response['list'];
-
-            foreach ($weather_list as $items) {
-                array_push($weathers, Location::getTranslation($items['weather'][0]['description'])); // 天気
-                array_push($temp_maxs, $items['main']['temp_max']); // 最高気温
-                array_push($temp_mins, $items['main']['temp_min']); // 最低気温
-                array_push($weather_icons, $items['weather'][0]['icon']); // 天気マーク
-                $datetime = new DateTime();
-                $datetime->setTimestamp($items['dt'])->setTimeZone(new DateTimeZone('Asia/Tokyo')); // 日時 - 協定世界時 (UTC)を日本標準時 (JST)に変換
-                array_push($dates, $datetime->format('Y年m月d日')); // 日付
-                array_push($times, $datetime->format('H:i')); // 時間
-            }
-
-            return view('locations.show', compact('temp_maxs', 'user', 'weathers', 'temp_mins', 'weather_icons', 'dates', 'times', 'tomorrow', 'location'));
+            return  $this->weatherShowService->getWeather($user, $location);
         } catch (Exception $e) {
-            // 例外発生時の処理
             $error[] = 'システムの都合上、この位置情報の天気を取得できません。大変申し訳ありません。';
             return redirect()->route('users.show', ['name' => $request->user()->name])->withInput()->withErrors($error);
         }
